type_struct
