+++ src/mpid/common/datatype/mpid_datatype_fallback.h
@@ -9,7 +9,6 @@
 #include "mpidu_datatype.h"
 
 /* MPID_Type */
-#define MPID_Type_struct MPIDU_Type_struct
 #define MPID_Type_vector MPIDU_Type_vector
 #define MPID_Type_zerolen MPIDU_Type_zerolen
